pyproject.toml
[tool.poetry]
name = "medicore"
version = "0.1.0"
description = ""
authors = ["betasecond <nbdhc9039@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
mysqlclient = "^2.2.7"
djangorestframework = "^3.15.2"


[[tool.poetry.source]]
name = "mirrors"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
